Categories x ja y arvot muutettu floatiksi koordinaattien täsmäämiseksi

+++ database/schema.sql
@@ -22,10 +22,10 @@ CREATE TABLE "Zone" (
 CREATE TABLE "Categories" (
   "category_id" int PRIMARY KEY,
   "name" varchar,
-  "x_min" int,
-  "x_max" int,
-  "y_min" int,
-  "y_max" int
+  "x_min" float,
+  "x_max" float,
+  "y_min" float,
+  "y_max" float
 );
 
 CREATE TABLE "ZoneVisit" (
